gateway/requirements.txt

requests
python-dotenv
colorama
elasticsearch
openai
aiohttp
fastmcp
mcp[cli]
